feat: add multiple users to list & log event
