+++ src/resolvers/fields/userList.ts
@@ -3,5 +3,6 @@ import resolver from '../resolver'
 export const total = resolver<number, any>()(
   async ({ knex, parent: { total, totalQuery } }) =>
     total ??
-    (await (totalQuery ?? knex('users')).count({ count: '*' }))[0].count
+    (await (totalQuery ?? knex('users')).count({ count: '*' }))[0]?.count ??
+    0
 )
