+++ src/resolvers/mutations/index.ts
@@ -2,7 +2,6 @@ export * from './account'
 export * from './profile'
 export * from './calendar'
 export * from './scheduling'
-export * from './list'
 export * from './picture'
 export * from './tag'
 export * from './invite'
